+++ src/Stack/Upload.hs
@@ -47,7 +47,7 @@ import           Data.Typeable                         (Typeable)
 import           Network.HTTP.Client                   (BodyReader, Manager,
                                                         Response,
                                                         RequestBody(RequestBodyLBS),
-                                                        applyBasicAuth, brRead,
+                                                        brRead,
                                                         newManager,
                                                         parseRequest,
                                                         requestHeaders,
@@ -55,7 +55,8 @@ import           Network.HTTP.Client                   (BodyReader, Manager,
                                                         responseStatus,
                                                         withResponse)
 import           Network.HTTP.Client.MultipartFormData (formDataBody, partFileRequestBody)
-import           Network.HTTP.Client.TLS               (tlsManagerSettings)
+import           Network.HTTP.Client.TLS               (tlsManagerSettings,
+                                                        applyDigestAuth)
 import           Network.HTTP.Types                    (statusCode)
 import           Path                                  (toFilePath)
 import           Prelude -- Fix redundant import warnings
@@ -203,10 +204,17 @@ mkUploader config us = do
         { upload_ = \tarName bytes -> do
             let formData = [partFileRequestBody "package" tarName (RequestBodyLBS bytes)]
             req2 <- formDataBody formData req1
-            let req3 = applyBasicAuth
+            mreq3 <- applyDigestAuth
                     (encodeUtf8 $ hcUsername creds)
                     (encodeUtf8 $ hcPassword creds)
                     req2
+                    manager
+            req3 <-
+                case mreq3 of
+                    Nothing -> do
+                        putStrLn $ "WARNING: No HTTP digest prompt found, this will probably fail"
+                        return req2
+                    Just req3 -> return req3
             putStr $ "Uploading " ++ tarName ++ "... "
             hFlush stdout
             withResponse req3 manager $ \res ->