+++ ChangeLog.md
@@ -16,6 +16,7 @@ Other enhancements:
 * `stack setup` detects systems where gcc enables PIE by default (such as Ubuntu
   16.10) and adjusts the GHC `configure` options accordingly.
   [#2542](https://github.com/commercialhaskell/stack/issues/2542)
+* Upload to Hackage with HTTP digest instead of HTTP basic.
 
 Bug fixes: